Option to switch between default and condensed display font profiles
On a 4:3 display, such as the ancient COLIN monitor, we might prefer the
old Roboto Condensed for the public display. So there is now an option
in the display setup to switch between Default and Condensed. This sets
a tourney option displayfontprofile to an integer which is an index into
countdowntourney.DISPLAY_FONT_PROFILES.

DISPLAY_FONT_PROFILES[x] gives the name of the font profile and the
basename of the CSS file under webroot/teleost/style/ to load before
webroot/teleost/style/main.css (which is the new name for
webroot/teleoststyle.css). That first CSS file contains font
definitions.