Show name/table index rather than fixture list for large tourneys
When new fixtures are generated and before the first result is entered,
the public display shows the fixture list. The organiser can tick a box
to say they want to show the name-to-table index at this point.

If a tournament has a large number of people (say, 36 or more players),
it's better to show the name-to-table index by default rather than the
fixture list, because it's more efficient. There are fewer pages to
display, so players wait less time to find out their table number.

The name-to-table index is therefore now the default for tourneys with
at least 36 active players. This can be overridden in the display
options page under the screen mode options for the "Automatic" screen
mode, as follows:

  * To switch off the name-to-table index entirely, uncheck the "Show
    name-to-table index..." checkbox.
  * To always show the name-to-table index at the start of a round
    regardless of the number of players, check the checkbox and set
    the "at least ... players" value to 0.
  * To show the name-to-table index at the start of a round only if you
    have more than N active players, check the checkbox and set the
    value to N. This is the default, where N = 36.

The name-to-table index now automatically defaults to a three-column
layout, which better fits a modern 16:9 display. To give long names the
best chance of fitting in the boxes, the name-to-table index uses a
condensed font (Roboto Condensed) if there are three or more columns.

On the display options page, the preview windows showing what the
public display is showing now have a 16:9 aspect ratio rather than 4:3.