Remove unused GAME_PENDING table and PendingPlayer
Long ago we had a "knockout" fixture generator, which could generate
fixtures whose players were not yet known but which depended on the
result of another game. For example, you could create a semi-final
whose players were "the winner of game X" and "the winner of game Y"
and the relevant player IDs for that game would be filled in
automatically when games X and Y were completed.

This hasn't been used or useable for years (possibly around the time
divisions were introduced?), and no fixture generators use it.
Knockout stages other than a final are now rare at co-events, but when
they do exist they're created manually one round at a time using the
Raw fixture generator. (Or, very often, the organiser simply doesn't
put the knockout stage into Atropine at all.)

So let's remove the GAME_PENDING table so it's no longer created in new
tourney databases, and remove the PlayerPending class, and remove all
the code that uses them. Also simplify any code that still checks if
the players for a game are "known" because this is now always the case.

The p1 and p2 fields of the GAME table are now defined as NOT NULL.

Get rid of generators/donotuse_fixgen_knockout.py.