"Discard..." link after fixture generation to return to main menu
After we generate fixtures, we give the user the choice between
accepting the fixtures and saving them to the database, or "Discard
and return to fixture generator". The "Discard..." option returns to
the specific fixture generator's menu, but this isn't very useful when
that fixture generator doesn't have a menu (such as the Round Robin
generator). In that case it just runs the fixture generator again and
presents the same fixtures again.

The "Discard..." link now goes back to the main fixture generator menu,
rather than the same fixture generator again.