Add Lucky Stiff to go with Tuff Luck
The Lucky Stiff prize exists in some Scrabble tournaments, and it is the
opposite of Tuff Luck. It's your aggregate winning margin over your
three closest wins. It is undefined if you won less than three games.

There's now a link in the sidebar for Lucky Stiff, and the Lucky Stiff
page looks a lot like the Tuff Luck page. It's almost as if it was
copied and adapted from it, or something.

countdowntourney.py's get_players_tuff_luck() method now returns the
number of defeats for each player as well as the other stuff.
Likewise, get_players_lucky_stiff() returns each player's number of
wins.