Improve and optimise SwissN fixture generator
Fix a few bugbears in the Swiss fixture generator and make it faster
by hitting it with the dynamic programming stick. This is apparently
what I spend an entire Sunday doing now.

First, the way it works out penalties has changed so that people on
"joint Nth" in the standings table are less likely to stick to each
other like glue to the detriment of people either side of them in the
standings table. For example, if we have people in 1st, 2nd, 3rd, 3rd,
3rd and 6th, beforehand it would put the 3rd-placed people on one table
and 1st, 2nd and 6th on another, much to the annoyance of the person
in 6th. Now the penalty given to a pair is calculated differently so
that it increases exponentially with their position difference, and
the penalty for 1st and 6th would outweigh the desirability of (3rd,
3rd, 3rd).

Second, it now runs much faster. This is because the recursive function
generate_all_groupings_aux() now caches the smaller sub-solutions so
that if we backtrack, rearrange the first few tables and call again
with the same remaining players, we already know the optimum
arrangement.

Also, redundant pieces of code are removed and refactored.
generate_sets() returns its results in a different order, to return
sets containing elements from a smaller sub-list of the ordered input
list first.

generate_all_groupings() and generate_all_groupings_aux() take
parameters with different data structures to make things easier.

If you have players with exactly the same standings position, these
players are now shuffled in a random order in the player list given
to swissN(), so that it's random which "joint Nth" player is nearer the
"top" and thus more likely to be picked first to go on a higher table.

Debug output from the swissN generator is reformatted.