Tweaks to Swiss fixture generator interface
Make the default "initial max win difference" 2 rather than 0. After the
SwissN optimisations, if there's a solution with a max win difference of
1 it'll probably find it anyway, but occasionally if we set the max win
difference to 0 or 1 then it can get preoccupied looking for a solution
with a max win difference of 1 and take 30 seconds to not find it.
(Reproduce with 22 players and 2 prunes.)

Use input type="number" rather than input type="text" for HTML controls
where that makes sense.