Script for setting up arbitrary scenarios and running swissN()
swissNscenariotest.py tests swissN.swissN() with a pre-supplied list
of games in test/swisstestscenarios/ and checks that the resulting
fixtures make sense and are within a specification.

Currently there are three scenario files in test/swisstestscenarios.
Each one supplies a list of players, a list of games, and a "test"
object containing desired properties of the fixtures swissN() generates
for the next round.