Bump version number to 1.2.0
The last Atropine release was 1.1.4 in March 2020. After a whole load of
new features and a two and a half year gap, it's time to bump the middle
component of the version number. The next tag and release will be
Atropine 1.2.0.

Previous version number was 1.1.5 but that was only bumped a couple of
weeks ago and never released except to a few people for testing.