Tweaks to fixpaths.sh script so it actually works on MacOS
sed on a Mac works slightly differently from how it works on Linux, so
the shell script needs to call it with different options.

This script is difficult to test because I don't own a Mac. Many thanks
to Mac owners Jack Hurst and Jen Steadman for helping to debug it.