Use HTML input type number instead of text where applicable
There are a few places where we use <input type="text" ...> when
<input type="number" ...> would be more appropriate. Change these to
type="number" so the user has those little arrows to click on to
change the number and, where useful, validation that the number is
within a prescribed range.

There are a few input tags which usually take numbers but are
deliberately being left as text. These are the rating field in a
player's info page (because we want to allow decimal places so we'd
have to say something like step=0.01 but that would make the up and
down arrows pretty useless), and the score inputs on the result entry
page (because we allow things like "30-57" and "66*" to be entered).

style.css: set the default width of all input tags of type "number" to
5em. The default otherwise is ridiculously long, at least in Chrome, and
this should remove the need for those 'style="width: 5em;"' attributes
peppered all over the place.

Add <label> tags to the division setup page where they're missing.