King of the Hill fixture generator should decide tied positions randomly
If two or more people share the same standings position (or, in round 1,
the same rating), those people are now put in a random order amongst
themselves in the list, so that each such player has an equal chance of
ending up on the stronger table. Previously, the person who was first
alphabetically ended up higher in the list.

Reword the KotH fixture generator's description.