Automatic Prunes - no more creating imaginary players with rating 0
If you don't have a multiple of three players but you're playing a
three-to-a-table event, until now you've had to add one or two Prune
players, with a rating of zero, to make up the numbers. This is
annoying, especially as you then have to remember to withdraw such a
Prune when another player joins late, or add/reinstate a Prune when a
player withdraws.

Let's introduce the Automatic Prune, a virtual player which turns up
only when fixture generators need it!

The PLAYER table in a newly-created tournament now comes with a row
already in it, with id -1, rating 0, and name "Prune".

This row is not deleted by resetting the player list, and nor does it
show up in any normal listing of players such as the standings table.
However, it can be referenced by the GAME table, and displayed as one of
a game's players. Fixture generators can thus decide how many Prunes are
needed based on the table size and the number of (real) players, and add
the Automatic Prune to as many tables as necessary.

Obviously all the fixture generators have been tweaked to support this.
Determining whether the database supports the automatic Prune, and if
so, which table(s) it sits at, is now the fixture generator's problem.

countdowntourney.get_auto_prune() returns a PrunePlayer, which extends
Player. The fixture generators use this object to represent a prune.
However, if the tourney database version is < 1.2.1, this throws an
exception to avoid trying to use the automatic prune on databases
without that extra player row. Fixture generators should call
tourney.has_auto_prune() to determine whether the tourney supports
automatic prunes.

Atropine version is now 1.2.1.

We don't need more than one automatic prune. If we're two short of a
multiple of three, we can just put the same Prune on two tables
rather than having two separate Prunes. The thing about a player whose
only job is to not exist and do nothing is that it's really good at
doing that in two places at once.

Various places that select * from the PLAYER table now have a "where id
>= 0" condition. ranktest.py has been tweaked to not expect Prune in the
standings, and not to manually add Prune to the player list. There are
a few other tweaks, which I won't further bulk out this already long
commit message to describe, to handle this strange new player which
appears in games but not in a standings table or player list.

Various other things which used to count Prune as a player now no longer
do so. This includes the name-to-table index and Tuff Luck. For the
purpose of the Tim Down Award, if you played the Automatic Prune, that
opponent is considered to have finished in last place, or joint last if
there are other prunes.

Finally, you can still create Prune players the "old" way, and this
works as before - create a player with a rating of zero. However, from
now on this is unlikely to be useful unless you need to create fixtures
on an old tourney database.