More auto-prune features, and ability to delete a player
The automatic prune can now be given a custom name in Advanced Setup.
This name must meet all the rules for player names - it may not be blank
or contain only whitespace, and it must not equal the name of any other
player.

A player who does not appear in any games can now be deleted from their
player page. Deleting the automatic prune is not permitted.

countdowntourney methods tweaked so that actions like
withdraw/reinstate, set rating, etc have no effect on the automatic
prune, not that there's any UI to do that anyway.

Reword the General Setup page to remove references to adding a Prune
player with a rating of zero, because you shouldn't need to do that any
more.