Refactor fixgen_manual, and don't allow a player on a table twice
fixgen_manual is majorly refactored with all the "how are these players
arranged, which slots contain duplicate players, which are empty, etc"
logic put in a PlayerArrangement class. This reads from the settings
object to determine which players are in which slots, and can be asked,
for each division, which slots are invalid in some way.

fixgen_manual now no longer lets you put the same player twice on the
same table. Before automatic prunes this was impossible anyway because
it wouldn't let you put the same player twice anywhere in the grid, but
we now allow the user to put more than one Prune in.

If any player, including Prune, appears more than once on the same row
of the grid (i.e. on the same table), treat this as a duplicate player
and treat the form as incomplete. We even do this if
allow_player_repetition is set, which the Raw fixture does. We don't
want to generate fixtures with players playing themselves.

Also: fixgen_swiss no longer warns the user that the player count isn't
a multiple of 2 or 3 if the tourney db supports auto-prunes.