Auto-fill Prune's score as 0 when entering result
When entering the result for a game involving a Prune, automatically
fill in "0" in the Prune's score, since that's almost always correct.

If you load a game's score into the result entry window by clicking
its blinkenlight or its entry in the table, then if one side is Prune
automatically focus and highlight the other score, rather than always
focusing and highlighting the first score.

The JavaScript dicts we build to help with auto-complete etc now use the
lowercase player name as the key, to remove the odd corner case where
this mattered. Everything should be completely case-insensitive now.