Display table numbers on a blue background everywhere
Wherever we display a table number, put it in a rounded rectangle with a
blue background where that makes sense. This applies in fixture lists,
on the result entry page, and in game lists on the player page.

We use black text on a pale blue background rather than the white on
saturated blue we see on the public display, because the web interface
notionally uses "light mode" while the display is "dark mode" and a
deep blue rectangle looks a bit out of place in the web interface.

The exception is the "blinkenlights" section of the result entry page,
which has always contained light text on a dark background so the table
numbers there are white on ordinary blue, or grey on navy blue if all
the games on that table have been played.

This makes things more consistent - the user knows without having to
check the label or heading that if a number is in a rounded rectangle
with a blue background, it's a table number.