Warn the user if they try to create a Prune manually
If the user enters a player with a rating of zero in the player list on
the General Setup page, the bit of the help that explains this is no
longer required now flashes yellow.

Also: if you submit a player list which contains a player whose name
equals the auto Prune name, we now throw a useful exception rather than
crashing with a uniqueness constraint check error from sqlite.