Display only bare minimum sidebar links if there are no players
When a tourney is initially created, pretty much the only thing you can
do is paste in and submit a player list. Let's remove most of the other
cruft from the sidebar so as not to distract from this.

tourneysetup.py is now heavily rearranged so that when the user submits
the player list, this is written to the database before we generate the
HTML for the sidebar, which needs to see the newly-submitted players so
it will give us the full set of links.

Also rearrange the controls in tourneysetup.py so that when the player
list is submitted, the "Tourney properties" section appears before the
player list controls. This is so the user doesn't miss them. Before the
player list is submitted, the "Tourney properties" section is not
displayed at all.