Show Welcome screen on display before any games are generated
Before any games are generated, previously we showed the standings
table. Everyone is on 0 wins and 0 points and in joint 1st at this
point, so the standings table is of almost no use except as a list of
players.

Instead, show a new "Welcome" view which displays a welcome message,
including the full name of the event if specified in General Setup.