Welcome screen to show list of active players
The Welcome screen now shows a list of all active players, in a table of
12 rows by 3 or 4 columns as appropriate. If there are more than 48
players we use two or more pages as necessary.

This gives players who have accidentally been omitted a chance to make
this known before the first games are generated.

WelcomeView is now a PagedTableView, very much like the
name-to-table-index view.

The name to table index view now shows "Page X of Y" in the corner if
there's more than one page, like the Welcome screen does.