Don't allow deletion of players who appear in games
countdowntourney.delete_player() must not delete a player if there is
at least one game featuring them. We aren't allowed to have rows in the
GAME table referencing nonexistent rows in the PLAYER table.

Perhaps this should be enforced in the database by a primary-foreign key
relationship but for some reason it's never been done like that. Perhaps
historically we couldn't rely on SQLite support for foreign keys?