Split Welcome screen into Welcome and Registration
Currently we've got a display mode which is effectively two modes. The
Welcome screen shows a welcome message if there are no players, and the
player registration table if there are players. Really this needs to be
two separate views, Welcome and Registration. Automatic mode should show
the Welcome view if there aren't players and the Registration view if
there are.

Give the Registration view a thumbnail containing unrivalled art.