Animate videprinter scrolling up before adding new entry
The videprinter view now animates itself scrolling up before animating
a new entry into the space at the bottom.

When redraw() is called for the first time after setup(), it just puts
the last four entries onto the videprinter without animating. When
redraw() is called subsequently, if there are any new unprinted entries
we first animate the videprinter table so it moves up over the course of
half a second, then (after a 0.5 second setTimeout()) add the new entry,
then when that animation has finished (we wait with another
setTimeout()), we do the redraw process again to check for any more
unprinted entries.

If a redraw is in progress waiting for one or other of the timeouts, an
external call to our redraw() has no effect. Our timeout handlers call
redrawInternal() which always goes through the redraw procedure.