Overhaul "Export results" functionality
Tidy up export.py. Put all the options into one form rather than having
four links, some of which lead to a form and some don't.

Remove the horrible old blue results tables and replace them with
something more like how they look on the results entry screen.

If there has been at least one final or third place match generated,
give the user a choice whether to export the standings table as it stood
before the finals, after the finals, or both. This is a vain attempt to
settle arguments about whether the standings table should be reordered
to take account of the result of the final or not.

"Download or display in browser" is now a choice you get with all four
file formats (HTML, Plain text, CSV, Wikitext) rather than just CSV.

standings.py now displays standings tables for both before and after the
finals, regardless of whether you checked the box in advanced setup to
apply the results of any finals to the standings table. This box now
only affects the public display and certain other things like what
position a player's individual page says they are in Stats Corner.

Tim Down Award: this is now based on the position of your opponents
before any finals, regardless of the state of that checkbox.