Validate date on export to Wikitext form
The Wikitext table prompts for the event date so it can be included in
the table as is customary on the Wiki. But the recent change to
export.py removed the validation on the date.

Validate the date in JavaScript whenever the day, month or year fields
are changed. If it's invalid, display a notice saying that the date is
invalid but if you insist on pressing the button we'll use it anyway.

If the day or year inputs are blank or contain something other than an
integer, then if you generate the Wikitext the date field of each game
is simply left blank.