Introduce vertically-split standings/results view
"Standings on the top, results on the bottom" is all very well for
the square displays of old, but most displays are widescreen now, and
the standings table doesn't need to be the full width of the screen.

On a widescreen display, it's a more efficient use of space to have the
standings table on the left half of the screen and the videprinter or
results on the right. Then we can show 12 standings positions at a time
instead of 8, and up to 18 games at a time instead of 3.

Occasionally, if someone has a very long name, it might exceed the width
allowed in the standings table or the results table, but we have the
"Condensed font" option in Display Setup for that case.

Introduce the new vertically-split standings/fixtures view, which can
optionally take the place of the horizontally-split standings/results
view and the horizontally-split standings/videprinter view. Options to
enable this are in Display Setup under the Auto view.

There is now code for a vertically-split standings/videprinter view but
nothing uses it yet.

Still to do: make a new icon for the vertically-split standings/fixtures
view, to appear on the Display Setup page. Currently it uses a
placeholder icon.