Improvements to Display Setup page, new thumbnails
The Display Setup page now has a better and less confusing layout.

The default "Auto" screen mode has a "suitable" thumbnail image, and
this is twice as large as the other thumbnails to emphasise that this
mode is the default and you probably want to leave it selected.

If you select a different mode and switch to it, a reminder "Auto OFF"
will appear in the sidebar until you re-enable Auto again.