Improvements to the "create tourney" page
The "create tourney" form now has some more stuff on it. The user can
now optionally specify the tourney's full name alongside the required
filename, and set the widescreen/not-widescreen option.

When the user presses the button to create a tourney, if successful, we
now redirect automatically to the new tourney's setup page rather than
making the user click a link to do that.

The Player Check-In link in the sidebar is now a bit more prominent.