Bump version number to 1.2.2 and improve tourney create redirect page
When you create a tourney, the page you see for a split second that says
"click this link if the redirect doesn't work" is now a plain page with
just that line of text on it, rather than an interesting-looking success
message box. Anything that looks interesting might lead the user to think
that they missed something important if they see it for a fraction of a
second before being redirected to the tourney setup page.