Divisioned Vertical Standings/Results screen can show wrong round
Division A: has a 7-round Round Robin, and rounds 1 and 2 are complete.
Division B: rounds 1 and 2 are complete but there's no round 3 yet.

The Vertical Standings/Results screen shows round 3 for Division A and
round 2 for Division B. It should show round 2 for both because Round 3
doesn't have any games yet.

The findSelectedRoundsPerDivision() function used by the fixtures view
now has a new mode (whichRnd == -5) which ensures the Vertical
Standings/Results screen shows the latest round which has any completed
games, and the same round for each division.

As a special case, if this means one of the divisions shows a round with
no games in it for that division, for that division we show the last
round before then which has at least one game.

The magic constant values of whichRnd (-1 to -5) now have symbolic
constants associated them so we no longer have to remember what -3 and
-5 actually mean when passed as whichRnd.