Tourney.get_current_round() can incorrectly return None
Division A: has a 7-round Round Robin. We've played rounds 1 and 2.
Division B: has rounds 1 and 2 complete, but no round 3.

Then we enter a result for a Division A game in round 3, before we
generate round 3 fixtures for Division B. The public display switches to
the Checkin screen because get_current_round() is returning None.

Fix: if the latest started round doesn't have games for every division
and this means it can't be the current round, return the latest suitable
round before that one rather than just returning None.