Introduce display options profiles
The Display Setup page now has a section at the bottom which lets you
save all the currently-applying display settings to a named profile.
These profiles are saved in prefs.db in the Atropine installation root,
which means they're available to all tourneys which use that Atropine
installation.

So if, for example, you want the Auto display mode to always announce
fixtures with the Fixtures view rather than the Table Number Index view
regardless of the number of players, you can now uncheck the relevant
box, tweak whatever other settings you like, save them as a named
profile, and then the next time you create a tourney, all you need to
do is load that profile back in.

The "Switch to profile" (load) button also has an option to revert back
to the "factory settings" for if you've messed everything up and just
want to go back to the defaults.

The Display Setup page allows you to save, load and delete profiles. All
settings on the Display Setup page, including the per-display-mode
settings, are saved to the profile, except the banner text and the
information about which mode is currently selected.

dialog.py: code to return HTML for a modal dialog box, for when the user
has pressed a button and we want to ask a question or check that they're
sure they want to do the thing.

The existing "global preferences" code, which used prefs.db and until
now has only stored one setting (game result entry tab order) is moved
from cgicommon.py to countdowntourney.py.