Allow user to choose display profile at tourney creation time
If there is at least one display profile defined in prefs.db, the
"create tourney" section of the home page now asks which display
profile you want to use for the tourney you're creating. The default
selection for this drop-down box is the display profile you last created
a tourney with.