Remove dependency on the deprecated cgi module
Python's cgi module is deprecated as of Python 3.11 and is being
removed in 3.13. Every Python CGI script in Atropine uses the
cgi.FieldStorage code to parse the script's query string and input
data, so we have to use an alternative.

Most of Atropine's CGI scripts use only the FieldStorage.getfirst()
method, but occasionally they use other functionality like the "in"
operator to test the presence of a field name.

webroot/cgi-bin/fieldstorage.py now contains a replacement FieldStorage
implementation. It doesn't implement all the functionality of
cgi.FieldStorage, only the subset used by Atropine. In particular, it
can only understand the input format "application/x-url-form-urlencoded"
and not "multipart/form-data", but that isn't a problem for Atropine
because the latter is only used for file upload forms and Atropine has
none of those.

The cgicommon module imports the replacement FieldStorage class from
fieldstorage.py. Every file that used to "import cgi" now doesn't, and
uses cgicommon.FieldStorage instead of cgi.FieldStorage.

Quite a few scripts that did "import cgi" didn't actually use the
cgi module anyway; in these cases the import statement is perhaps a
fossil from when we used the cgi.escape() function, which is now
html.escape().