Remove dependency on cgitb module
Like the cgi module, the cgitb module is also deprecated from Python
3.11 and will be removed in Python 3.13. So let's introduce an
alternative, htmltraceback.py, which is less good but at least won't
be randomly taken away in the future.

It works the same way as cgitb does, by setting sys.excepthook to a
custom function. sys.excepthook is called immediately before Python dies
of an uncaught exception.

Call htmltraceback.enable() from a CGI script to enable it, in place of
cgitb.enable().