Players can be marked as newbies
Add a "newbie" column to the PLAYER table, and make the Player setup
page include a checkbox to display and change a player's newbie status.
Atropine does not currently use this information for anything, but the
intention is that the Random fixture generator can later be configured
to avoid generating all-newbie tables in the first round.

To set a player's newbie status in the player list on the Tourney
Setup page, add a field "N" after a player's name. For example:

    Purple Trouserpress,N

means that Purple Trouserpress is a newbie.

countdowntourney.py: instead of comparing db_version to magic
constants to test whether the database we're looking at has the
relevant column for a specific feature, and replace them with methods
like has_player_newbie_feature() instead, so all the logic depending
on db_version is in the same place.

The Atropine software version number is bumped to 1.2.3. Tourney
database files earlier than that are not expected to have a newbie
column in the PLAYER table.

Other housekeeping: the code to fetch player attributes from the player
table, and substitute default values for column names if the database
doesn't have those columns, is repeated in multiple places. This logic
is now refactored into get_players_where().

If a tourney database file is too old to have columns for certain
player attributes, don't show controls for them on the Player Setup
page. This applies to the newbie field (absent pre-1.2.3) and some
other controls like whether the player should avoid playing Prune or
whether they require an accessible table.