Random fixture generators can now enforce no all-newbie tables
The Random and Random from Seeded Pots fixture generators can now
enforce that every table has at least one non-newbie on it. The option
is only presented if there is at least one player marked as a newbie,
and if the option is presented, it defaults to on.

In order to ensure no all-newbie tables, there must be at least as
many non-newbies as there are tables. If this isn't the case, the
fixture generator will notice this and give an error.

Prunes count as newbies for fixture generation - if we have a table of
two newbies and one Prune, we have to count it as an all-newbie table.

The Random fixture generator could already avoid rematches. However,
there is no facility for it to both avoid rematches *and* avoid
all-newbie tables. You get a radio button asking if you want to avoid
rematches, or avoid all-newbie tables, or neither. The "both"
combination is not likely to be required anyway; if rematches are
possible then we've already played at least one round, so it's less
important to have a non-newbie on every table.

The "avoid rematches" option is now only presented if there are already
games in the tourney database.

randomfixgentest.py: some tests for the Random and Random from Seeded
Pots fixture generators. Some test cases just ensure the fixture
generators run, don't throw exceptions on valid input, and return an
acceptable set of fixtures. A couple of them run the fixture generator a
large number of times to make sure we get the correct number of distinct
combinations - in other words, that the "random" fixture generator is
capable of returning every acceptable combination.