Deduplicate dialog box code
The Player Check-In page has its own code to display dialog boxes, which
existed before the dialog.py module. This page now uses the dialog.py
module, which has been extensively modified to accommodate the features
required by the Player Check-In page, the Display Setup page, and the
Player page.

The dialog.py module now has a get_html() function which returns the
HTML for an initially-invisible div element which can be placed anywhere
on the page. There is also a DIALOG_JAVASCRIPT string which the page
must put in a <script> tag somewhere on the page. To show the dialog
box, call the JavaScript function dialogBoxShow(), defined in
DIALOG_JAVASCRIPT. This function must be passed the dialog box's
title, button labels, any hidden form elements to be posted with the
submit button, and the additional DOM elements to put into the form.

The Player page now shows an "are you sure" dialog box when you press
the button to delete a player. That button is now no longer bright red.

Code and CSS definitions for the old check-in dialog box are removed.