Introduce new random-drawing engine
py/randomdraw.py provides a function draw() which will put the numbers
in the range [0, n), representing players, on tables of the specified
size(s). It takes a list invalid_pairs, which is a list of player
numbers (x, y). For each pair (x, y), x and y will not be in the same
returned group.

The draw() function first puts the given players on tables at random
and checks if that fits the given constraints. If not, it tries again
a set number of times, and if it still hasn't found an acceptable
grouping, it does a systematic but random search for an acceptable
grouping.

The systematic search, if needed, does not guarantee that every valid
outcome is equally likely (a 50/50 decision might lead to 10 solutions
down one path but only one solution down another, so the solution on
the latter path is therefore 10 times more likely than each individual
solution on the former) but it's better than giving up or taking ages
to find a valid grouping in cases where there are a lot of constraints.

The random fixture generator, and potentially some future fixture
generators such as "each person plays a random opponent on the same
number of wins", can use this code instead of what it does now, which
can struggle if there are a large number of players but not many valid
grouping outcomes.