Replace jsonreq.py CGI script with subclassed HTTP handler
cgi-bin/jsonreq.py is gone. This was used by the public-facing display
and the sidebar uploader widget. Its replacement is the URL
/atropine/<tourneyname>/state/ which can be followed by "default" or
"uploader" or anything else which was always accepted by
tourney2json.py.

All HTTP requests now go through AtropineHTTPRequestHandler, defined in
atropine.py. This is a subclass of CGIHTTPRequestHandler. If the path
part of the URL begins with /atropine/, then the next path component is
expected to be the tourney name. The path component after that is used
to look up which handler in py/httphandler/ will handle the request.

So if the browser requests /atropine/mytourney/state/default, we open
mytourney.db, and call the handle() function in
py/httphandler/http_handler_state.py. This function replies to the
browser.

The result is that the regular requests by the public-facing display and
the uploader widget no longer use CGI and therefore no longer spawn a
new process. In time, more things might be able to work this way, for
example, we might be able to submit a score without refreshing the page.