Tell browser not to cache content returned from /atropine/<tourney>/...
The browser might not realise that a URL with a path like
/atropine/mytourney/state/default is dynamic content, because it
doesn't have a query string in it. Of course, the response contains
standings and scores which might change at any time, so the browser
shouldn't cache the response - when any script on a page requests that
URL, the browser should actually request and fetch it.

handlerutils.send_response() now has a cache_max_sec parameter which
defaults to 0. The response includes a Cache-Control header which tells
the browser that it may cache the result for this many seconds, or if 0,
that it may not cache it at all.

main_post.js: fix the URL requested by the public-facing display to
remove the unnecessary ampersand, and remove the question mark if
there's no query string.