Add option to Swiss generator to match players only by wins
Implement the much requested "play random opponent(s) on the same number
of wins as you" draw by adding an option to the Swiss Army Blunderbuss
generator.

When enabled, the Swiss generator will disregard the players' standings
positions when deciding the penalty for matching up two given players.
This penalty will be based only on the difference in wins between the
two players. If the players are on the same number of wins and have not
already played each other, the penalty for that pair of players is zero.
Otherwise, if two players are on different numbers of wins, there is (a)
an exponentially-increasing penalty based on the win count difference,
plus (b) a penalty for how far the lower-win-count player is below the
top of their win group in the standings.

Players on the same win count have their order in the initial list
randomised before we search for groupings.

The effect of all this is that we try to match players with opponents on
the same win count ((a) above), but it's random which specific
opponent(s) on the same win count are matched up. Often the number of
players on a certain number of wins will not be a multiple of the table
size; if we must choose someone to play an opponent on a higher win
count, we prefer to choose the highest-ranked player with the lower
win count ((b) above).

A hill-climb stage is introduced after the main "generate a load of
groupings and see which is best" stage. This is because the new mode
would often come up with groupings that could be trivially improved by
swapping two players from different tables, so that's what we do. See if
we can swap two players to improve the overall penalty, and keep doing
that until this doesn't improve it any more.